test(core): split cache module coverage
